made search tab and fragment
