+++ SousBox/app/src/main/java/com/billy/sousbox/sousbox/adapters/RecyclerClicker/ClickListener.java
@@ -1,0 +1,15 @@
+package com.billy.sousbox.sousbox.adapters.RecyclerClicker;
+
+import android.view.View;
+
+/**
+ * Created by Billy on 5/22/16.
+ */
+public interface ClickListener {
+
+    void onClick(View view, int position);
+
+    void onLongClick(View view, int position);
+}
+
+
