+++ SousBox/app/src/main/java/com/billy/sousbox/sousbox/adapters/RecyclerClicker/RecyclerTouchListener.java
@@ -1,0 +1,53 @@
+package com.billy.sousbox.sousbox.adapters.RecyclerClicker;
+
+import android.content.Context;
+import android.support.v7.widget.RecyclerView;
+import android.view.GestureDetector;
+import android.view.MotionEvent;
+import android.view.View;
+
+/**
+ * Created by Billy on 5/22/16.
+ */
+public class RecyclerTouchListener implements RecyclerView.OnItemTouchListener {
+
+private GestureDetector gestureDetector;
+private ClickListener clickListener;
+
+public RecyclerTouchListener(Context context, final RecyclerView recyclerView, final ClickListener clickListener) {
+        this.clickListener = clickListener;
+        gestureDetector = new GestureDetector(context, new GestureDetector.SimpleOnGestureListener() {
+@Override
+public boolean onSingleTapUp(MotionEvent e) {
+        return true;
+        }
+
+@Override
+public void onLongPress(MotionEvent e) {
+        View child = recyclerView.findChildViewUnder(e.getX(), e.getY());
+        if (child != null && clickListener != null) {
+        clickListener.onLongClick(child, recyclerView.getChildPosition(child));
+        }
+        }
+        });
+        }
+
+@Override
+public boolean onInterceptTouchEvent(RecyclerView rv, MotionEvent e) {
+        View child = rv.findChildViewUnder(e.getX(), e.getY());
+        if (child != null && clickListener != null && gestureDetector.onTouchEvent(e)) {
+        clickListener.onClick(child, rv.getChildPosition(child));
+        }
+        return false;
+        }
+
+@Override
+public void onTouchEvent(RecyclerView rv, MotionEvent e) {
+        }
+
+@Override
+public void onRequestDisallowInterceptTouchEvent(boolean disallowIntercept) {
+
+        }
+
+}
