+++ SousBox/app/src/main/java/com/billy/sousbox/sousbox/api/RecipeAPI.java
@@ -17,7 +17,7 @@ public interface RecipeAPI {
 
     //delete for final version
     @Headers("X-Mashape-Key: " + Keys.MASHAPLE)
-    @GET("search?limitLicense=false&number=75&offset=0&")
+    @GET("search?limitLicense=false&number=50&offset=0&")
     Call<SpoonacularResults> searchRecipe(@Query("query")String q);
 
     @Headers("X-Mashape-Key: " + Keys.MASHAPLE)
